requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
deprecation==2.1.0
exceptiongroup==1.2.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.0
gotrue==2.4.1
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.25.2
idna==3.6
lxml==5.1.0
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
postgrest==0.16.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
realtime==1.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.18.1
sentry-sdk==2.3.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
storage3==0.7.3
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
tqdm==4.66.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
webdriver-manager==4.0.2
websockets==11.0.3
wsproto==1.2.0
